// firestore/functions/src/main/Https/tasks/seedSubtasks/interface.ts
export interface seedSubtaskInterface {
  id: string;
  taskName: string;
  kcal: number;
  protein: number;
  carbs: number;
  fats: number;
  fiber: number;
  weight_nouse: number;
  serving_unit: string;
  serving_weight: number;
  kcal_100g: number;
  protein_100g: number;
  carbs_100g: number;
  fats_100g: number;
  fiber_100g: number;
}